+++ tests/test_slice.py
@@ -9,41 +9,70 @@ import xarray as xr
 
 from zappend.context import Context
 from zappend.fsutil.fileobj import FileObj
-from zappend.slicesource.common import open_slice_source
-from zappend.slicesource.memory import MemorySliceSource
-from zappend.slicesource.persistent import PersistentSliceSource
-from zappend.slicesource.temporary import TemporarySliceSource
+from zappend.slice.common import get_slice_dataset
+from zappend.slice.memory import MemorySliceSource
+from zappend.slice.persistent import PersistentSliceSource
+from zappend.slice.temporary import TemporarySliceSource
 from .helpers import clear_memory_fs
 from .helpers import make_test_dataset
 
 
-class SliceSourceTest(unittest.TestCase):
+class OpenSliceDatasetTest(unittest.TestCase):
     def setUp(self):
         clear_memory_fs()
+
+    def test_slice_source_slice_source(self):
+        dataset = make_test_dataset()
+        ctx = Context(dict(target_dir="memory://target.zarr"))
+        slice_obj = MemorySliceSource(ctx, dataset, 0)
+        slice_source = get_slice_dataset(ctx, slice_obj)
+        self.assertIs(slice_obj, slice_source)
+
+    def test_factory_slice_source(self):
+        dataset = make_test_dataset()
+        ctx = Context(dict(target_dir="memory://target.zarr"))
+
+        def factory(_ctx):
+            self.assertIs(ctx, _ctx)
+            return dataset
+
+        slice_source = get_slice_dataset(ctx, factory)
+        self.assertIsInstance(slice_source, MemorySliceSource)
+        with slice_source as slice_ds:
+            self.assertIs(dataset, slice_ds)
 
     def test_memory_slice_source(self):
         dataset = make_test_dataset()
         ctx = Context(dict(target_dir="memory://target.zarr"))
-        slice_zarr = open_slice_source(ctx, dataset)
-        self.assertIsInstance(slice_zarr, MemorySliceSource)
-        with slice_zarr as slice_ds:
+        slice_source = get_slice_dataset(ctx, dataset)
+        self.assertIsInstance(slice_source, MemorySliceSource)
+        with slice_source as slice_ds:
             self.assertIsInstance(slice_ds, xr.Dataset)
 
     def test_temporary_slice_source(self):
         dataset = make_test_dataset()
         ctx = Context(dict(target_dir="memory://target.zarr", persist_mem_slices=True))
-        slice_zarr = open_slice_source(ctx, dataset)
-        self.assertIsInstance(slice_zarr, TemporarySliceSource)
-        with slice_zarr as slice_ds:
+        slice_source = get_slice_dataset(ctx, dataset)
+        self.assertIsInstance(slice_source, TemporarySliceSource)
+        with slice_source as slice_ds:
+            self.assertIsInstance(slice_ds, xr.Dataset)
+
+    def test_file_obj_slice_source(self):
+        slice_dir = FileObj("memory://slice.zarr")
+        make_test_dataset(uri=slice_dir.uri)
+        ctx = Context(dict(target_dir="memory://target.zarr"))
+        slice_source = get_slice_dataset(ctx, slice_dir)
+        self.assertIsInstance(slice_source, PersistentSliceSource)
+        with slice_source as slice_ds:
             self.assertIsInstance(slice_ds, xr.Dataset)
 
     def test_persistent_slice_source_for_zarr(self):
         slice_dir = FileObj("memory://slice.zarr")
         make_test_dataset(uri=slice_dir.uri)
         ctx = Context(dict(target_dir="memory://target.zarr"))
-        slice_zarr = open_slice_source(ctx, slice_dir.uri)
-        self.assertIsInstance(slice_zarr, PersistentSliceSource)
-        with slice_zarr as slice_ds:
+        slice_source = get_slice_dataset(ctx, slice_dir.uri)
+        self.assertIsInstance(slice_source, PersistentSliceSource)
+        with slice_source as slice_ds:
             self.assertIsInstance(slice_ds, xr.Dataset)
 
     # def test_persistent_slice_source_for_nc(self):
@@ -67,13 +96,13 @@ class SliceSourceTest(unittest.TestCase):
                 slice_polling=dict(timeout=0.1, interval=0.02),
             )
         )
-        slice_zarr = open_slice_source(ctx, slice_dir.uri)
-        self.assertIsInstance(slice_zarr, PersistentSliceSource)
-        with slice_zarr as slice_ds:
+        slice_source = get_slice_dataset(ctx, slice_dir.uri)
+        self.assertIsInstance(slice_source, PersistentSliceSource)
+        with slice_source as slice_ds:
             self.assertIsInstance(slice_ds, xr.Dataset)
 
     # noinspection PyMethodMayBeStatic
-    def test_persistent_wait_fail(self):
+    def test_it_raises_on_persistent_wait_fail(self):
         slice_dir = FileObj("memory://slice.zarr")
         ctx = Context(
             dict(
@@ -81,18 +110,14 @@ class SliceSourceTest(unittest.TestCase):
                 slice_polling=dict(timeout=0.1, interval=0.02),
             )
         )
-        slice_zarr = open_slice_source(ctx, slice_dir.uri)
+        slice_source = get_slice_dataset(ctx, slice_dir.uri)
         with pytest.raises(FileNotFoundError, match=slice_dir.uri):
-            with slice_zarr:
+            with slice_source:
                 pass
 
     # noinspection PyMethodMayBeStatic
-    def test_invalid_type(self):
-        dataset_dir = FileObj("memory://slice.zarr")
-        make_test_dataset(uri=dataset_dir.uri)
+    def test_it_raises_on_invalid_type(self):
         ctx = Context(dict(target_dir="memory://target.zarr"))
-        with pytest.raises(
-            TypeError, match="slice_obj must be a str or xarray.Dataset"
-        ):
+        with pytest.raises(TypeError, match="slice_obj must be a str, "):
             # noinspection PyTypeChecker
-            open_slice_source(ctx, dataset_dir)
+            get_slice_dataset(ctx, 42)